Further refactoring of statistical tests into individual tests
