+++ tests/PhpSpreadsheetTests/Calculation/Functions/Statistical/BetaDistTest.php
@@ -1,0 +1,31 @@
+<?php
+
+namespace PhpOffice\PhpSpreadsheetTests\Calculation\Functions\Statistical;
+
+use PhpOffice\PhpSpreadsheet\Calculation\Functions;
+use PhpOffice\PhpSpreadsheet\Calculation\Statistical;
+use PHPUnit\Framework\TestCase;
+
+class BetaDistTest extends TestCase
+{
+    public function setUp()
+    {
+        Functions::setCompatibilityMode(Functions::COMPATIBILITY_EXCEL);
+    }
+
+    /**
+     * @dataProvider providerBETADIST
+     *
+     * @param mixed $expectedResult
+     */
+    public function testBETADIST($expectedResult, ...$args)
+    {
+        $result = Statistical::BETADIST(...$args);
+        self::assertEquals($expectedResult, $result, '', 1E-12);
+    }
+
+    public function providerBETADIST()
+    {
+        return require 'data/Calculation/Statistical/BETADIST.php';
+    }
+}
