+++ tests/PhpSpreadsheetTests/Calculation/Functions/Statistical/MaxIfsTest.php
@@ -1,0 +1,31 @@
+<?php
+
+namespace PhpOffice\PhpSpreadsheetTests\Calculation\Functions\Statistical;
+
+use PhpOffice\PhpSpreadsheet\Calculation\Functions;
+use PhpOffice\PhpSpreadsheet\Calculation\Statistical;
+use PHPUnit\Framework\TestCase;
+
+class MaxIfsTest extends TestCase
+{
+    public function setUp()
+    {
+        Functions::setCompatibilityMode(Functions::COMPATIBILITY_EXCEL);
+    }
+
+    /**
+     * @dataProvider providerMAXIFS
+     *
+     * @param mixed $expectedResult
+     */
+    public function testMAXIFS($expectedResult, ...$args)
+    {
+        $result = Statistical::MAXIFS(...$args);
+        self::assertEquals($expectedResult, $result, '', 1E-12);
+    }
+
+    public function providerMAXIFS()
+    {
+        return require 'data/Calculation/Statistical/MAXIFS.php';
+    }
+}
