+++ tests/PhpSpreadsheetTests/Calculation/Functions/Statistical/MinIfsTest.php
@@ -1,0 +1,31 @@
+<?php
+
+namespace PhpOffice\PhpSpreadsheetTests\Calculation\Functions\Statistical;
+
+use PhpOffice\PhpSpreadsheet\Calculation\Functions;
+use PhpOffice\PhpSpreadsheet\Calculation\Statistical;
+use PHPUnit\Framework\TestCase;
+
+class MinIfsTest extends TestCase
+{
+    public function setUp()
+    {
+        Functions::setCompatibilityMode(Functions::COMPATIBILITY_EXCEL);
+    }
+
+    /**
+     * @dataProvider providerMINIFS
+     *
+     * @param mixed $expectedResult
+     */
+    public function testMINIFS($expectedResult, ...$args)
+    {
+        $result = Statistical::MINIFS(...$args);
+        self::assertEquals($expectedResult, $result, '', 1E-12);
+    }
+
+    public function providerMINIFS()
+    {
+        return require 'data/Calculation/Statistical/MINIFS.php';
+    }
+}
