+++ tests/data/Calculation/Statistical/CONFIDENCE.php
@@ -5,4 +5,12 @@ return [
         0.692951912734,
         0.05, 2.5, 50,
     ],
+    [
+        0.48999099653,
+        0.05, 2.5, 100,
+    ],
+    [
+        0.013719747903,
+        0.05, 0.07, 100,
+    ],
 ];
